+++ lib/widgets/talawa_error_dialog.dart
@@ -1,17 +1,25 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class TalawaErrorWidget extends StatelessWidget {
-  const TalawaErrorWidget(this.errorMessage, {Key? key}) : super(key: key);
+class TalawaErrorDialog extends StatelessWidget {
+  const TalawaErrorDialog(this.errorMessage, {Key? key}) : super(key: key);
   final String errorMessage;
   @override
   Widget build(BuildContext context) {
     return SizedBox(
       child: AlertDialog(
-        title: const Text("Error", style: TextStyle(color: Colors.red)),
-        content: SingleChildScrollView(
-          child: ListBody(
-            children: <Widget>[Text(errorMessage)],
+        title: const Text(
+          "Error",
+          style: TextStyle(color: Colors.red),
+        ),
+        content: SizedBox(
+          width: 200,
+          height: 75,
+          child: AutoSizeText(
+            errorMessage,
+            style: const TextStyle(fontSize: 20),
+            maxLines: 3,
           ),
         ),
         actions: <Widget>[
@@ -26,34 +34,3 @@ class TalawaErrorWidget extends StatelessWidget {
     );
   }
 }
-
-// void showTalawaErrorDialog(String errorMessage) {
-//   showDialog(
-//     context: navigatorKey.currentContext!,
-//     barrierColor: Colors.transparent,
-//     barrierDismissible: false,
-//     builder: (BuildContext context) {
-//       return SizedBox(
-//         child: AlertDialog(
-//           title: const Text(
-//             "Error",
-//             style: TextStyle(color: Colors.red),
-//           ),
-//           content: SingleChildScrollView(
-//             child: ListBody(
-//               children: <Widget>[Text(errorMessage)],
-//             ),
-//           ),
-//           actions: <Widget>[
-//             TextButton(
-//               child: const Text('Close'),
-//               onPressed: () {
-//                 Navigator.of(context).pop();
-//               },
-//             ),
-//           ],
-//         ),
-//       );
-//     },
-//   );
-// }
